# specs/001-performance-service/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Performance Management Service API
  description: |
    API for managing employee performance reviews, goals, performance improvement plans (PIPs), and 360-degree feedback.

    ## Authentication
    All endpoints require JWT authentication via Bearer token in the `Authorization` header.

    ## Permissions
    - `performance.create`: Create performance reviews
    - `performance.read`: View reviews and goals
    - `performance.update`: Update reviews and goals
    - `performance.admin`: Manage PIPs
    - `performance.feedback`: Submit feedback

    ## Base Path
    All endpoints are prefixed with `/performance/v1`
  version: 1.0.0
  contact:
    name: MALIEV Co., Ltd.
    email: api@maliev.com
servers:
  - url: https://api.maliev.com/performance/v1
    description: Production
  - url: https://staging-api.maliev.com/performance/v1
    description: Staging
  - url: http://localhost:8080/performance/v1
    description: Local development
...
paths:
...
  /employees/{employeeId}/goals:
    get:
      tags:
        - Goals
      summary: Get goals for an employee
      description: Retrieves all goals for a specific employee with optional filtering
      operationId: getGoals
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/GoalStatus'
          description: Filter by goal status
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedGoals'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: [performance.read]

    post:
      tags:
        - Goals
      summary: Create a new goal
      description: Creates a new goal for an employee, optionally linked to a review period
      operationId: createGoal
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalRequest'
      responses:
        '201':
          description: Goal created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - BearerAuth: [performance.update]
  /goals/{goalId}:
    put:
      tags:
        - Goals
      summary: Update goal
      description: Updates goal description, success criteria, or target date
      operationId: updateGoal
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGoalRequest'
      responses:
        '200':
          description: Goal updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: [performance.update]
  /goals/{goalId}/progress:
    put:
      tags:
        - Goals
      summary: Update goal progress
      description: Adds a progress update to the goal and optionally updates status
      operationId: updateGoalProgress
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGoalProgressRequest'
      responses:
        '200':
          description: Progress updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: [performance.update]
  /employees/{employeeId}/pips:
    get:
      tags:
        - PIPs
      summary: Get PIPs for an employee
      description: Retrieves all Performance Improvement Plans for a specific employee
      operationId: getPIPs
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PIP'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: [performance.read]

    post:
      tags:
        - PIPs
      summary: Create a Performance Improvement Plan
      description: Initiates a new PIP for an employee (HR/Admin only)
      operationId: createPIP
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePIPRequest'
      responses:
        '201':
          description: PIP created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PIP'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Employee already has an active PIP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: PIP_ALREADY_ACTIVE
                message: Employee already has an active PIP
                timestamp: 2025-12-28T10:00:00Z
      security:
        - BearerAuth: [performance.admin]
  /pips/{pipId}/outcome:
    put:
      tags:
        - PIPs
      summary: Record PIP outcome
      description: Records the outcome of a completed PIP (HR/Admin only)
      operationId: recordPIPOutcome
      parameters:
        - $ref: '#/components/parameters/PIPId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordPIPOutcomeRequest'
      responses:
        '200':
          description: PIP outcome recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PIP'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: [performance.admin]
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from authentication service
  parameters:
    EmployeeId:
      name: employeeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Employee UUID
...
    GoalId:
      name: goalId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Goal UUID
    PIPId:
      name: pipId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: PIP UUID
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Pagination cursor (base64 encoded timestamp|id)
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
  responses:
    Unauthorized:
      description: Unauthorized - Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: Missing or invalid authentication token
            timestamp: 2025-12-28T10:00:00Z
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_AUTHORIZED
            message: Insufficient permissions to perform this action
            timestamp: 2025-12-28T10:00:00Z
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: REVIEW_NOT_FOUND
            message: Performance review not found
            timestamp: 2025-12-28T10:00:00Z
    BadRequest:
      description: Bad Request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
...
    GoalStatus:
      type: string
      enum: [NotStarted, InProgress, AtRisk, Completed, Deferred, Cancelled]
    PIPStatus:
      type: string
      enum: [Active, Extended, Completed, Terminated]
    PIPOutcome:
      type: string
      enum: [Successful, Unsuccessful, ExtendedAgain]
...
    Goal:
      type: object
      required:
        - id
        - employeeId
        - description
        - targetDate
        - completionStatus
        - createdDate
      properties:
        id:
          type: string
          format: uuid
        employeeId:
          type: string
          format: uuid
        performanceReviewId:
          type: string
          format: uuid
          nullable: true
        description:
          type: string
        successCriteria:
          type: string
          nullable: true
        targetDate:
          type: string
          format: date-time
        completionStatus:
          $ref: '#/components/schemas/GoalStatus'
        progressUpdates:
          type: string
          nullable: true
          description: Append-only progress notes with timestamps
        completedDate:
          type: string
          format: date-time
          nullable: true
        createdDate:
          type: string
          format: date-time
        modifiedDate:
          type: string
          format: date-time
          nullable: true
    CreateGoalRequest:
      type: object
      required:
        - description
        - targetDate
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 1000
        successCriteria:
          type: string
          nullable: true
        targetDate:
          type: string
          format: date-time
        performanceReviewId:
          type: string
          format: uuid
          nullable: true
    UpdateGoalRequest:
      type: object
      properties:
        description:
          type: string
        successCriteria:
          type: string
        targetDate:
          type: string
          format: date-time
    UpdateGoalProgressRequest:
      type: object
      required:
        - progressUpdate
      properties:
        progressUpdate:
          type: string
          minLength: 1
          maxLength: 500
        completionStatus:
          $ref: '#/components/schemas/GoalStatus'
    PagedGoals:
      type: object
      required:
        - items
        - hasNextPage
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Goal'
        nextCursor:
          type: string
          nullable: true
        hasNextPage:
          type: boolean
    PIP:
      type: object
      required:
        - id
        - employeeId
        - initiatedBy
        - startDate
        - endDate
        - reason
        - improvementAreas
        - successCriteria
        - status
        - createdDate
      properties:
        id:
          type: string
          format: uuid
        employeeId:
          type: string
          format: uuid
        initiatedBy:
          type: string
          format: uuid
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        reason:
          type: string
        improvementAreas:
          type: string
        successCriteria:
          type: string
        checkInNotes:
          type: string
          nullable: true
          description: Append-only check-in notes with timestamps
        status:
          $ref: '#/components/schemas/PIPStatus'
        outcome:
          $ref: '#/components/schemas/PIPOutcome'
          nullable: true
        createdDate:
          type: string
          format: date-time
        modifiedDate:
          type: string
          format: date-time
          nullable: true
    CreatePIPRequest:
      type: object
      required:
        - startDate
        - endDate
        - reason
        - improvementAreas
        - successCriteria
      properties:
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        reason:
          type: string
          minLength: 1
        improvementAreas:
          type: string
          minLength: 1
        successCriteria:
          type: string
          minLength: 1
    RecordPIPOutcomeRequest:
      type: object
      required:
        - outcome
      properties:
        outcome:
          $ref: '#/components/schemas/PIPOutcome'
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
    ValidationErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
        - validationErrors
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: One or more validation errors occurred
        timestamp:
          type: string
          format: date-time
        validationErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          example:
            - field: reviewPeriodStart
              message: Review period start date must be before end date